requirements.txt
urllib3==1.26.15    # specify a macos openssl compatible version of urllib3
langchain-ollama
argparse
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
python-dotenv
